+++ goldens/circular-deps/packages.json
@@ -1003,37 +1003,13 @@
     "packages/core/testing/src/test_bed.ts"
   ],
   [
-    "packages/forms/src/directives/abstract_control_directive.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts"
-  ],
-  [
-    "packages/forms/src/directives/abstract_control_directive.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/directives/abstract_form_group_directive.ts",
-    "packages/forms/src/directives/control_container.ts"
-  ],
-  [
-    "packages/forms/src/directives/abstract_control_directive.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/directives/abstract_form_group_directive.ts",
-    "packages/forms/src/directives/control_container.ts",
-    "packages/forms/src/directives/form_interface.ts",
-    "packages/forms/src/directives/ng_control.ts"
-  ],
-  [
     "packages/forms/src/directives/abstract_form_group_directive.ts",
     "packages/forms/src/directives/control_container.ts",
     "packages/forms/src/directives/form_interface.ts"
   ],
   [
     "packages/forms/src/directives/abstract_form_group_directive.ts",
-    "packages/forms/src/directives/control_container.ts",
-    "packages/forms/src/directives/form_interface.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts"
+    "packages/forms/src/directives/form_interface.ts"
   ],
   [
     "packages/forms/src/directives/abstract_form_group_directive.ts",
@@ -1041,13 +1017,59 @@
   ],
   [
     "packages/forms/src/directives/abstract_form_group_directive.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts"
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/control_container.ts",
+    "packages/forms/src/directives/form_interface.ts"
   ],
   [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/ng_control.ts",
     "packages/forms/src/directives/control_container.ts",
-    "packages/forms/src/directives/form_interface.ts",
-    "packages/forms/src/directives/ng_control.ts"
+    "packages/forms/src/directives/form_interface.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
+    "packages/forms/src/directives/control_container.ts",
+    "packages/forms/src/directives/form_interface.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
+    "packages/forms/src/directives/control_container.ts",
+    "packages/forms/src/directives/form_interface.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
+    "packages/forms/src/directives/form_interface.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
+    "packages/forms/src/directives/reactive_directives/form_control_name.ts"
+  ],
+  [
+    "packages/forms/src/directives/abstract_form_group_directive.ts",
+    "packages/forms/src/directives/shared.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
+    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
+    "packages/forms/src/directives/reactive_directives/form_control_name.ts",
+    "packages/forms/src/directives/control_container.ts",
+    "packages/forms/src/directives/form_interface.ts"
   ],
   [
     "packages/forms/src/directives/ng_form.ts",
@@ -1066,14 +1088,6 @@
     "packages/forms/src/directives/reactive_directives/form_control_name.ts"
   ],
   [
-    "packages/forms/src/directives/reactive_directives/form_control_directive.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
-    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
-    "packages/forms/src/directives/reactive_directives/form_control_name.ts"
-  ],
-  [
     "packages/forms/src/directives/reactive_directives/form_control_name.ts",
     "packages/forms/src/directives/reactive_directives/form_group_directive.ts"
   ],
@@ -1084,13 +1098,6 @@
   ],
   [
     "packages/forms/src/directives/reactive_directives/form_control_name.ts",
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
-    "packages/forms/src/directives/reactive_directives/form_group_directive.ts"
-  ],
-  [
-    "packages/forms/src/directives/reactive_directives/form_control_name.ts",
-    "packages/forms/src/model.ts",
     "packages/forms/src/directives/shared.ts",
     "packages/forms/src/directives/reactive_directives/form_group_name.ts",
     "packages/forms/src/directives/reactive_directives/form_group_directive.ts"
@@ -1101,12 +1108,6 @@
   ],
   [
     "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/directives/reactive_directives/form_group_name.ts"
-  ],
-  [
-    "packages/forms/src/directives/reactive_directives/form_group_directive.ts",
-    "packages/forms/src/model.ts",
     "packages/forms/src/directives/shared.ts",
     "packages/forms/src/directives/reactive_directives/form_group_name.ts"
   ],
@@ -1115,28 +1116,17 @@
     "packages/forms/src/directives/shared.ts"
   ],
   [
-    "packages/forms/src/directives/reactive_directives/form_group_name.ts",
-    "packages/forms/src/model.ts",
-    "packages/forms/src/directives/shared.ts"
-  ],
-  [
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/model.ts"
-  ],
-  [
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/validators.ts",
     "packages/forms/src/directives/validators.ts",
-    "packages/forms/src/model.ts"
-  ],
-  [
-    "packages/forms/src/directives/shared.ts",
-    "packages/forms/src/validators.ts",
     "packages/forms/src/model.ts"
   ],
   [
     "packages/forms/src/directives/validators.ts",
     "packages/forms/src/validators.ts"
+  ],
+  [
+    "packages/forms/src/directives/validators.ts",
+    "packages/forms/src/validators.ts",
+    "packages/forms/src/model.ts"
   ],
   [
     "packages/language-service/src/completions.ts",
